adding benchmarks and generated new land.pkl
